Added configuration script
+ The configuration script (./config.sh) is used to set the root
    of the website.
* Renamed app/index.html to app/raw-index.html and replaced the base
    value with $ROOT (which gets replaced by the config value).
* Renamed app/build/app.build.js to app/build/raw-app.build.js and
    replaced the output value with $DEST. (replaced by another config
    value)
* Configuration settings are saved to ./.config.  For info on what each
    line represents, refer to the config.sh file which writes them
    in that order.
* Changed the build script (./build.sh) to use the config settings and
    build the scss files.

+++ build.sh
@@ -1,9 +1,36 @@
-rm -rf dist
+# First, check for a config
+if [ ! -f .config ]
+  then echo "Run config.sh first!"
+  exit
+fi
+
+ROOT=$(sed '1q;d' .config)
+DEST=$(sed '2q;d' .config)
+
+rm -rf ${DEST}
 rm -rf app/data
+
+echo "compile scss"
+cd app
+compass compile
+cd ..
+
+echo "init data"
 mkdir app/data
 cp app/rawdata/ability.poke app/data/ability.json
 rhino pokemon.js > app/data/pokemon.json
 rhino move.js > app/data/move.json
+
+echo "fix options"
+cp app/raw-index.html app/index.html
+sed -i "s/\$ROOT/$(echo ${ROOT} | sed 's/\//\\\//g')/g" app/index.html
+cp app/build/raw-app.build.js app/build/app.build.js
+sed -i "s/\$DEST/$(echo ${DEST} | sed 's/\//\\\//g')/g" app/build/app.build.js
+
+echo "compiling..."
 r.js -o app/build/app.build.js
-cd dist
+
+echo "cleaning up"
+cd ${DEST}
+#TODO: Add more things to delete from ${DEST} here.
 rm -rf build build.txt js/views js/vendor bower_components js/foundation js/backbone.localStorage-min.js js/foundation.min.js js/models js/collections js/views js/router.js fbower_components sass config.rb rawdata
